backend/requirements.txt
alembic==1.8.1
aniso8601==9.0.1
attrs==22.1.0
click==8.1.3
colorama==0.4.5
Flask==1.1.2
Flask-Cors==3.0.9
Flask-HTTPAuth==4.7.0
Flask-Migrate==2.6.0
Flask-RESTful==0.3.8
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==4.16.0
Mako==1.2.3
MarkupSafe==2.0.1
pkgutil-resolve-name==1.3.10
PyJWT==2.5.0
pyrsistent==0.18.1
python-dotenv==0.21.0
pytz==2022.2.1
six==1.16.0
SQLAlchemy==1.3.22
Werkzeug==1.0.1
zipp==3.8.1
